requirements.txt
pytorch-lightning
lightning-bolts
torchmetrics
rtpt
wandb
yapf
pre-commit
sklearn
torchattacks
hydra-core
omegaconf
scikit-learn